move all tests into core that are related to asciidoc2Confluence
