+++ core/build.gradle
@@ -15,4 +15,16 @@ dependencies {
     implementation((libs.http.builder))
     implementation((libs.httpmime))
     implementation((libs.jsoup))
+
+    testImplementation(libs.spock.reports) {
+        exclude group: 'org.codehaus.groovy', module: 'groovy-xml'
+    }
+    testImplementation(
+        (libs.junit),
+        (libs.groovy.all),
+    )
+}
+
+test {
+    useJUnitPlatform()
 }
