+++ core/src/test/resources/asciidoc2confluence/fixtures/body.html
