+++ core/src/test/resources/asciidoc2confluence/fixtures/imageBase64.txt
