+++ core/src/test/resources/asciidoc2confluence/images/diagram_classes.png
